some modification added in course

+++ StudentManagementSystem/Services/CourseService.cs
@@ -16,20 +16,38 @@ namespace StudentManagementSystem.Services
         private IRepository<Course> courseRepository;
         private IRepository<LocalizedProperty> localPropertyRepository;
         private IRepository<Language> languageRepository;
+        private IRepository<StudentCourse> studentCourseRepository;
+        private IRepository<CourseTeacher> CourseTeacherRepository;
         private const string CourseTable = "Course";
         private const string CourseName = "Name";
         private const string CourseCredit = "Credit";
         private const string BaseLanguage = "English";
         public CourseService(IRepository<Course> courseRepository, IRepository<LocalizedProperty> localPropertyRepository,
-            IRepository<Language> languageRepository)
+            IRepository<Language> languageRepository, IRepository<StudentCourse> studentCourseRepository,
+            IRepository<CourseTeacher> courseTeacherRepository)
         {
             this.courseRepository = courseRepository;
             this.localPropertyRepository = localPropertyRepository;
             this.languageRepository = languageRepository;
+            this.studentCourseRepository = studentCourseRepository;
+            this.CourseTeacherRepository = courseTeacherRepository;
         }
-        public void DeleteCourse(int id)
+        public bool DeleteCourse(int id)
         {
-            courseRepository.Delete(id);
+            var studentcourses = studentCourseRepository.GetAll().Where(sc => sc.CourseId == id).FirstOrDefault();
+            var courseTeachers = CourseTeacherRepository.GetAll().Where(ct => ct.CourseId == id).FirstOrDefault();
+
+            if (studentcourses == null || courseTeachers == null)
+            {
+                courseRepository.Delete(id);
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+
+            
         }
 
         public Course GetCourse(int id)
